+++ clang-tools-extra/docs/ReleaseNotes.rst
@@ -255,6 +255,11 @@ New checks
   Finds calls to STL library iterator algorithms that could be replaced with
   LLVM range-based algorithms from ``llvm/ADT/STLExtras.h``.
 
+- New :doc:`misc-anonymous-namespace-in-header
+  <clang-tidy/checks/misc/anonymous-namespace-in-header>` check.
+
+  FIXME: Write a short description.
+
 - New :doc:`misc-override-with-different-visibility
   <clang-tidy/checks/misc/override-with-different-visibility>` check.